Add NY info to the output
This will add NYC and NY State data to the bot. NY State DOH does
not provide negative case data, so instead the delineation is
between NYC and the rest of the state. Anywhere not in the city
is considered "upstate". Anywhere on Long Island east of BK and
Queens is "upstate". This is the state of things.

We should probably split all this duped code into a function, but
this is quick and dirty with unwashed hands.

Signed-off-by: Cullen Taylor <cullen@eggshell.me>